+++ CHANGELOG.md
@@ -3,6 +3,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
  * Search index skips updates when task timestamp predates index document.
+ * Reduced frequency and concurrency of search snapshot writes to storage bucket.
 
 ## `20200619t091224-all`
  * Bumped runtimeVersion to `2020.06.19`.